work in progress

+++ app/src/main/java/com/example/kinopoisk/api/kinopoisk/response/FilmSearchResponse.kt
@@ -19,7 +19,7 @@ data class FilmSearchResponse(
         @SerializedName("description")
         val description: String,
         @SerializedName("filmId")
-        val filmId: Long,
+        val filmId: Int,
         @SerializedName("filmLength")
         val filmLength: String,
         @SerializedName("genres")
